+++ examples/blockly-parcel/src/index.js
@@ -11,6 +11,7 @@
  */
 
 import * as Blockly from 'blockly';
+import {javascriptGenerator} from 'blockly/javascript';
 
 const toolbox = {
   kind: 'flyoutToolbox',
@@ -88,11 +89,10 @@ document.addEventListener('DOMContentLoaded', function () {
       media: 'media/'
     });
 
-  const lang = 'JavaScript';
   const button = document.getElementById('blocklyButton');
   button.addEventListener('click', function () {
     alert('Check the console for the generated output.');
-    const code = Blockly[lang].workspaceToCode(workspace);
+    const code = javascriptGenerator.workspaceToCode(workspace);
     console.log(code);
   });
 });