+++ plugins/field-date/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.0.4](https://github.com/google/blockly-samples/compare/@blockly/field-date@6.0.3...@blockly/field-date@6.0.4) (2023-01-05)
+
+**Note:** Version bump only for package @blockly/field-date
+
+
+
+
+
+## [6.0.3](https://github.com/google/blockly-samples/compare/@blockly/field-date@6.0.2...@blockly/field-date@6.0.3) (2022-12-15)
+
+**Note:** Version bump only for package @blockly/field-date
+
+
+
+
+
 ## [6.0.2](https://github.com/google/blockly-samples/compare/@blockly/field-date@6.0.1...@blockly/field-date@6.0.2) (2022-10-27)
 
 **Note:** Version bump only for package @blockly/field-date