+++ plugins/theme-deuteranopia/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.2](https://github.com/google/blockly-samples/compare/@blockly/theme-deuteranopia@3.0.1...@blockly/theme-deuteranopia@3.0.2) (2023-01-05)
+
+**Note:** Version bump only for package @blockly/theme-deuteranopia
+
+
+
+
+
 ## [3.0.1](https://github.com/google/blockly-samples/compare/@blockly/theme-deuteranopia@3.0.0...@blockly/theme-deuteranopia@3.0.1) (2022-10-13)
 
 **Note:** Version bump only for package @blockly/theme-deuteranopia